final VimL config commit
Changing the plugin path so that I can use stow for the nvim config.
This will be the final viml config. I'm in the process of porting over
to Lua.
It's currently in a mess but I should figure out the best practices and
what's most convenient the more changes I make